Updated Code to control from single scripts | QOL updates

+++ blockchain/scripts/remove.sh
@@ -1,25 +1,18 @@
 echo "Removing all Containers of Healthchain Blockchain"
-docker rm -f healthchain_cli.blc.healthchain.com_1
-docker rm -f healthchain_orderer.blc.healthchain.com_1
-docker rm -f healthchain_ca.id1.national.healthchain.com_1
-docker rm -f healthchain_peer0.id2.local.healthchain.com_1
-docker rm -f healthchain_peer0.id1.national.healthchain.com_1
-docker rm -f healthchain_ca.id1.local.healthchain.com_1
-docker rm -f healthchain_peer0.id1.local.healthchain.com_1
-docker rm -f healthchain_ca.id1.state.healthchain.com_1
-docker rm -f healthchain_peer0.id1.state.healthchain.com_1
-docker rm -f healthchain_ca.id2.local.healthchain.com_1
-
-docker rm -f healthchain_api.blc.id1.national.healthchain.com_1
-docker rm -f healthchain_api.blc.id1.state.healthchain.com_1
-docker rm -f healthchain_api.blc.id1.local.healthchain.com_1
-docker rm -f healthchain_api.blc.id2.local.healthchain.com_1
-
-# remove docker volumes
-docker volume ls --format '{{.Name}}' | grep healthchain.com | xargs docker volume rm -f
-
-# remove the chaincode containers and images
+docker ps -a --format '{{.Names}}' | grep healthchain_ca | xargs docker rm -f
+docker ps -a --format '{{.Names}}' | grep healthchain_peer0 | xargs docker rm -f
+docker ps -a --format '{{.Names}}' | grep healthchain_orderer | xargs docker rm -f
+docker ps -a --format '{{.Names}}' | grep healthchain_cli | xargs docker rm -f
+docker ps -a --format '{{.Names}}' | grep healthchain_api.blc | xargs docker rm -f
 docker ps -a --format '{{.Names}}' | grep dev-peer | xargs docker rm -f
-docker images --format '{{.Repository}}' | grep dev-peer | xargs docker image rm -f
 
-docker images --format '{{.Repository}}' | grep api.blc | xargs docker image rm -f
+# remove the docker images create previously
+if [[ "$1" = "-c" ]]; then
+    echo "Removing Images, Volumes and Networks as well...."
+    # remove docker volumes
+    docker volume ls --format '{{.Name}}' | grep healthchain.com | xargs docker volume rm -f
+
+    # remove the chaincode images
+    docker images --format '{{.Repository}}' | grep dev-peer | xargs docker image rm -f
+    docker images --format '{{.Repository}}' | grep api.blc | xargs docker image rm -f
+fi
